.ipynb_checkpoints/requirements-checkpoint.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-openai
langchain-chroma
chromadb

langchain-text-splitters
pydantic
